feat: more reliable form of call service start / stop (based on current call)
