+++ telnyx_common/src/main/java/com/telnyx/webrtc/common/TelnyxCommon.kt
@@ -3,10 +3,14 @@ package com.telnyx.webrtc.common
 import android.content.Context
 import android.content.SharedPreferences
 import androidx.lifecycle.Observer
+import com.telnyx.webrtc.common.service.CallForegroundService
 import com.telnyx.webrtc.sdk.Call
 import com.telnyx.webrtc.sdk.TelnyxClient
+import com.telnyx.webrtc.sdk.model.PushMetaData
+import com.telnyx.webrtc.sdk.verto.receive.InviteResponse
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
+import timber.log.Timber
 import java.util.*
 
 /**
@@ -53,18 +57,22 @@ class TelnyxCommon private constructor() {
         }
     }
 
-    internal fun setCurrentCall(call: Call?) {
+    internal fun setCurrentCall(context: Context, call: Call?) {
         call?.let { newCall ->
             telnyxClient?.getActiveCalls()?.get(newCall.callId)?.let {
                 _currentCall = it
+                // Start the CallForegroundService
+                startCallService(context, it)
             }
         } ?: run {
             _currentCall = null
+            // if we have no active call, stop the CallForegroundService
+            stopCallService(context)
         }
     }
 
     internal fun registerCall(call: Call) {
-        val holdStatusObserver = Observer<Boolean> { value ->
+        val holdStatusObserver = Observer<Boolean> { _ ->
             updateHoldedCalls()
         }
         holdStatusObservers[call] = holdStatusObserver
@@ -74,6 +82,41 @@ class TelnyxCommon private constructor() {
     internal fun unregisterCall(callId: UUID) {
         holdStatusObservers.entries.find { it.key.callId == callId }?.let { entry ->
             entry.key.getIsOnHoldStatus().removeObserver(entry.value)
+        }
+    }
+
+    private fun startCallService(viewContext: Context, call: Call?) {
+        // Check if the service is already running
+        if (CallForegroundService.isServiceRunning(viewContext)) {
+            Timber.d("CallForegroundService is already running, not starting again")
+            return
+        }
+
+        val pushMetaData = PushMetaData(
+            callerName = call?.inviteResponse?.callerIdName ?: "Active Call",
+            callerNumber =  call?.inviteResponse?.callerIdNumber ?: "",
+            callId =  call?.callId.toString(),
+        )
+
+        try {
+            // Start the foreground service
+            CallForegroundService.startService(viewContext, pushMetaData)
+            Timber.d("Started CallForegroundService for ongoing call")
+        } catch (e: IllegalStateException) {
+            Timber.e(e, "Failed to start CallForegroundService: ${e.message}")
+        }
+    }
+
+    private fun stopCallService(context: Context) {
+        try {
+            context.let {
+                if (CallForegroundService.isServiceRunning(it)) {
+                    CallForegroundService.stopService(it)
+                    Timber.d("Stopped CallForegroundService after call ended by remote party")
+                }
+            }
+        } catch (e: IllegalStateException) {
+            Timber.e(e, "Failed to stop CallForegroundService: ${e.message}")
         }
     }
