+++ telnyx_common/src/main/java/com/telnyx/webrtc/common/domain/call/OnByeReceived.kt
@@ -1,5 +1,6 @@
 package com.telnyx.webrtc.common.domain.call
 
+import android.content.Context
 import com.telnyx.webrtc.common.TelnyxCommon
 import java.util.*
 
@@ -14,10 +15,10 @@ class OnByeReceived {
      *
      * @param callId The unique identifier of the call to end.
      */
-    operator fun invoke(callId: UUID) {
+    operator fun invoke(context: Context, callId: UUID) {
         val telnyxCommon = TelnyxCommon.getInstance()
         if (telnyxCommon.currentCall?.callId == callId)
-            telnyxCommon.setCurrentCall(null)
+            telnyxCommon.setCurrentCall(context, null)
 
         telnyxCommon.unregisterCall(callId)
     }
